+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v2.3.0
+
+* Added the following fields to response objects (@shivashankar-ror):
+  - CaptureResponse: `:respproc, :resptext, :commcard, :respstat, :respcode, :batchid, :token`
+  - InquireResponse: `:capturedate, :batchid, :token, :authdate, :lastfour, :name, :currency, :settledate`
+  - ProfileGetResponse: `:gsacard, :auoptout`
+
+* Added support for Bin Service (@shivashankar-ror)
+
+## v2.2.0
+
+* Add `token` field to Profile Service for create/update responses since this field can potentially be included in the response if a token hadn't already been created for that particular profile. (@mauriciozaffari)
+
 ## v2.1.0
 
 * Add ability to pass connection options to the Faraday connection (@jhanggi)